flask_project/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
advertorch==0.2.3
fvcore==0.1.5.post20220512
matplotlib==3.4.2
numpy
tqdm==4.61.0
scipy
Flask-Executor